+++ config/me_cms.php
@@ -22,6 +22,8 @@ return ['MeCms' => [
 	'frontend' => [
 		//Google Analytics ID or FALSE
 		'analytics' => FALSE,
+		//It shows the alert for the cookie policy
+		'cookies_policy' => TRUE,
 		//Layout you want to use for the frontend
 		//If you want to use a layout from your application (eg. `default.ctp`), change this value without extension
 		'layout' => 'MeCms.frontend',